+++ botx/testing.py
@@ -21,6 +21,7 @@ APIMessage = Union[
     requests.UpdatePayload,
     requests.StealthEnablePayload,
     requests.StealthDisablePayload,
+    requests.AddRemoveUsersPayload,
 ]
 
 
@@ -125,6 +126,16 @@ class _BotXAPICallbacksFactory:
 
         return factory
 
+    def get_add_remove_user_callback(self) -> Callable:
+        """Generate callback for add/remove user endpoints."""  # noqa: D202
+
+        async def factory(request: Request) -> JSONResponse:
+            payload = requests.AddRemoveUsersPayload.parse_obj(await request.json())
+            self.messages.append(payload)
+            return JSONResponse(responses.AddRemoveUserResponse().dict())
+
+        return factory
+
 
 def _botx_api_mock(
     messages: List[APIMessage], generate_errored: bool = False
@@ -169,6 +180,16 @@ def _botx_api_mock(
         BotXAPI.stealth_disable_endpoint.endpoint,
         factory.get_stealth_disable_callback(),
         [BotXAPI.stealth_disable_endpoint.method],
+    )
+    app.add_route(
+        BotXAPI.add_user_endpoint.endpoint,
+        factory.get_add_remove_user_callback(),
+        [BotXAPI.add_user_endpoint.method],
+    )
+    app.add_route(
+        BotXAPI.remove_user_endpoint.endpoint,
+        factory.get_add_remove_user_callback(),
+        [BotXAPI.remove_user_endpoint.method],
     )
     return app