+++ docs/changelog.md
@@ -14,6 +14,7 @@
 
 * Refactored next step middleware
 * Next step middleware won't now lookup for handler in bot.
+* Disable `loguru` logger by default.
 
 ## 0.13.2 (Feb 14, 2020)