+++ docs/changelog.md
@@ -2,7 +2,20 @@
 
 ### Added
 
-* Methods to add/remove users to/from chat `Bot.add_users()` and `Bot.remove_users()`.
+* Add background dependencies to next step middleware.
+* Next step break handler can be registered as function.
+* Add methods to add/remove users to/from chat using `Bot.add_users()` and `Bot.remove_users()`.
+
+### Fixed
+
+* Add missing `dependency_overrides_provider` to `botx.collecting.Collector.add_handler`.
+* Encode message update payload by alias.
+
+### Changed
+
+* Refactored next step middleware
+* Next step middleware won't now lookup for handler in bot.
+* Disable `loguru` logger by default.
 
 ## 0.13.2 (Feb 14, 2020)